requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
cssselect2==0.8.0
Django==5.2.11
django-environ==0.12.0
djangorestframework==3.16.1
flake8==7.3.0
freetype-py==2.5.1
html5lib==1.1
idna==3.11
isort==7.0.0
lxml==6.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
oscrypto==1.3.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
pycairo==1.29.0
pycodestyle==2.14.0
pycparser==3.0
pycryptodome==3.23.0
pyflakes==3.4.0
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.6.2
python-bidi==0.6.7
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
xhtml2pdf==0.2.17
